re-organize backend server file to match VxScan
